requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

click == 8.0.3
httpx == 0.22.0
isodate == 0.6.1
loguru == 0.6.0
lxml == 4.7.1
protobuf == 3.19.4
pycryptodomex == 3.14.0
rich == 11.1.0
typer == 0.4.0
